fix dashboard agora comparando resultados

+++ serverr/src/main/java/com/ecommerce/serverr/dto/DashboardDTO.java
@@ -2,7 +2,6 @@ package com.ecommerce.serverr.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -13,6 +12,6 @@ import java.time.LocalDate;
 public class DashboardDTO {
     private String titulo;
     private Double valorTotalPedidoItem;
-    private BigDecimal itemQuantidade;
+    private Integer itemQuantidade;
     private LocalDate dataPedido;
 }
